// Frontend/src/constants/enums.ts
export enum ButtonVariants {
  CONTAINED = "contained",
  OUTLINED = "outlined",
  TEXT = "text",
}

export enum InputVariants {
  OUTLINED = "outlined",
  FILLED = "filled",
  STANDARD = "standard",
}

export enum ToastAlertColor {
  SUCCESS = "success",
  ERROR = "error",
  INFO = "info",
  WARNING = "warning",
}

export enum ColumnAlignment {
  LEFT = "left",
  RIGHT = "right",
  CENTER = "center",
}

export enum SortOrder {
  ASC = "asc",
  DESC = "desc",
}

export enum CruxMetrics {
  lcp = "lcp",
  fid = "fid",
  cls = "cls",
  fcp = "fcp",
  ttfb = "ttfb",
  INP = "INP",
}

export enum Rating {
  good = "good",
  imprv = "needs-improvement",
  poor = "poor",
}

export enum FormFactor {
  PHONE = "PHONE",
  DESKTOP = "DESKTOP",
  TABLET = "TABLET",
}